feat: Added Flow Sample Type
The Flow Sample has different data elements than the Counter Sample type. Each of these need to be parsed differently for the Sample and Record level. The initial testing shows that both Samples are working properly but the records are not. The Flow Records appear to be off-set by 4 bytes.