fix: Enabled detection of struct 1-8800-2
The Brocade vRouter is using the PMACCT code from https://github.com/pmacct/pmacctwhich is generating an undocumented struct, 1-8800-2. The code was not identifying the Enterprise for the Struct when parsing.

The code has been temporarily fixed by adding a test for the Enterprise for struct 1-0-2. This logic needs to be applied to all structures as there may be other unidentified or expected structs as additional systems are tested agianst the code.

THe Test display has been reduced to only display unknown structs and their enterprise.